% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_control.R
\name{gam_control}
\alias{gam_control}
\title{Auxiliary function for controlling the GAM copula fitting}
\usage{
gam_control(
  formula = ~1,
  family_set = c(1, 2, 301:304, 401:404),
  tau = TRUE,
  method = "FS",
  tol.rel = 0.001,
  n.iters = 10,
  ...
)
}
\arguments{
\item{formula}{a formula object, starting with the \code{~} operator and
followed by the covariates, separated by \code{+} operators.
Default: \code{~ 1}, i.e. only intercept.}

\item{family_set}{a vector of bivariate copula families as implemented in
\code{\link{gamCopula}}, i.e. \code{1} Gaussian, \code{2} Student-t,
\code{301} Double Clayton type I (standard and rotated 90 degrees),
\code{302} Double Clayton type II (standard and rotated 270 degrees),
\code{303} Double Clayton type III (survival and rotated 90 degrees),
\code{304} Double Clayton type IV (survival and rotated 270 degrees),
\code{401} Double Gumbel type I (standard and rotated 90 degrees),
\code{402} Double Gumbel type II (standard and rotated 270 degrees),
\code{403} Double Gumbel type III (survival and rotated 90 degrees),
\code{404} Double Gumbel type IV (survival and rotated 270 degrees).
All available copula families are used per default.}

\item{tau}{logical; specifies whether Kendall's tau (default: \code{TRUE}) or the
copula parameter (\code{FALSE}) should be parametrized by the \code{formula}.}

\item{method}{character; Fisher-scoring (default: \code{"FS"}) or Newton-Raphson
(default: \code{"NR"}) for the fitting.}

\item{tol.rel}{double; relative tolerance for \code{"FS"/"NR"} algorithm.
Default: \code{tol.rel = 0.001}.}

\item{n.iters}{integer; maximal number of iterations for \code{"FS"/"NR"} algorithm.
Default: \code{n.iters = 10}.}

\item{...}{unused.}
}
\value{
a list with components named as the arguments.
}
\description{
Auxiliary function for controlling the GAM copula fitting
}
