% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamvinereg.R
\name{gamvinereg}
\alias{gamvinereg}
\title{D-vine GAM copula based quantile regression}
\usage{
gamvinereg(
  formula,
  data,
  margins,
  selcrit = "AIC",
  control = gam_control(),
  order = NA,
  uscale = FALSE,
  cores = 1,
  ...
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of an
\code{~} operator, and the covariates, separated by \code{+} operators,
on the right for the D-vine.}

\item{data}{data frame containing all variables for the D-vine GAM copula.}

\item{margins}{list of margin specification(s) for each variable in the same order as in the formula argument.
Each margin distribution is specified as list by the arguments \code{"family"}, \code{"mu.formula"}, \code{"sigma.formula"}
and if necessary \code{"nu.formula"} and \code{"tau.formula"} according to \code{\link{gamlss}}.}

\item{selcrit}{selection criterion for the margins and GAM copulas.
\code{"AIC"} (default), \code{"BIC"} or \code{"logLik"} are the possible choices.}

\item{control}{control parameters for the GAM copulas. The \code{\link{gam_control}} function
provides the default setting.}

\item{order}{the order of covariates in the D-vine, provided as vector of
variable names.}

\item{uscale}{logical; if \code{TRUE}, then \code{gamvinereg} assumes that the variables for the D-vine
are already in [0,1] and the \code{margins} argument will be ignored.
Default setting is \code{uscale = FALSE}.}

\item{cores}{integer; the number of cores used for computations.
Default setting is \code{cores = 1}.}

\item{...}{unused.}
}
\value{
An object of class \code{gamvinereg}. It is a list containing the elements
\describe{
  \item{formula}{the formula used for the D-vine fit.}
  \item{gam_formula}{the formula used for the GAM copula fit.}
  \item{selcrit}{criterion used for margin and variable selection.}
  \item{model_frame}{the data used to fit the regression model.}
  \item{margins}{list consisting of the selected models, compared margin
  specifications and its corresponding statistics.}
  \item{vine}{a list consisting of the selected pair copulas, the D-vine
  strucuture and the variable names.}
  \item{stats}{fit statistics such as conditional
  log-likelihood/AIC/BIC and p-values for each variable's contribution.}
  \item{order}{order of the covariates chosen by the variable selection
  algorithm.}
  \item{selected_vars}{indices of selected variables.}}
Use the \code{\link{predict.gamvinereg}} function to predict conditional quantiles.
Furthermore, apply the \code{\link{cpit}} function to calculate the conditional
probability integral transform (PIT) values. Additionally, use the
\code{\link{cpdf}} function to calculate the conditional density values.
The \code{summary} function of a \code{gamvinereg} object shows the contribution of
each selected variable with the associated p-value derived from a likelihood ratio test.
}
\description{
D-vine GAM copula based quantile regression as described by Jobst, Möller and Groß (2023).
}
\examples{
# load data for station Hannover
data(station)

# formula for gamvinereg
formula <- obs ~ t2m_mean + t2m_sd

# data transformation to uniform scale
ustation <- as.data.frame(VineCopula::pobs(station[, all.vars(formula)]))

## generate margin distributions
# generate left-truncated at 0 normal distribution
gen.trun(par = 0, family = "NO", type = "left")
# generate left-censored normal distribution
gen.cens(family = "NO", type = "left")
# generate log-normal distribution
gen.Family(family = "NO", type = "log")

## margin specifications to select from
margins <- list(

 # margin for variable obs (normal distribution)
 list(list(family = NO(),
           mu.formula = obs ~ sin1 + cos1,
           sigma.formula = ~ sin1 + cos1)),

 # margin for variable t2m_mean (normal distribution)
 list(list(family = NO(),
           mu.formula = t2m_mean ~ sin1 + cos1,
           sigma.formula = ~ sin1 + cos1)),

 # margins for variable t2m_sd (log-normal, left-truncated/censored at 0 normal distribution)
 list(list(family = logNO(),
           mu.formula = t2m_sd ~ sin1 + cos1,
           sigma.formula = ~ sin1 + cos1),
      list(family = NOtr(),
           mu.formula = t2m_sd ~ sin1 + cos1,
           sigma.formula = ~ sin1 + cos1),
      list(family = NOlc(),
           mu.formula = Surv(t2m_sd, t2m_sd > 0, type = "left") ~ sin1 + cos1,
           sigma.formula = ~ sin1 + cos1))

)


# fit gamvinereg with time-dependent linear correlation model
(object1 <- gamvinereg(formula = formula,
                       data = station,
                       control = gam_control(formula = ~ sin1 + cos1),
                       margins = margins))

# fit gamvinereg with time-dependent spline based correlation model
(object2 <- gamvinereg(formula = formula,
                       data = station,
                       control = gam_control(formula = ~ s(doy, bs = "cc")),
                       margins = margins))

# fit gamvinereg with on uniform data with constant correlation model
# fixed variable order t2m_sd - t2m_mean
(object3 <- gamvinereg(formula = formula,
                       data = ustation,
                       order = c("t2m_sd", "t2m_mean"),
                       uscale = TRUE))

# predict median
med <- predict(object1, alpha = 0.5)

# predict deciles
dec <- predict(object2, alpha = 1:9/10)

# predict quartiles
q <- predict(object3, newdata = ustation, alpha = c(0.25, 0.75))

}
\references{
Kraus, D. and Czado, C. (2017). D-vine copula based quantile regression,
Computational Statistics and Data Analysis, 110, 1-18.doi: 10.1016/j.csda.2016.12.009.

Jobst, D., Möller, A., and Groß, J. (2023). D-Vine GAM Copula based
Quantile Regression with Application to Ensemble Postprocessing. doi: 10.48550/arXiv.1510.04161.
}
