% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\docType{data}
\name{station}
\alias{station}
\title{Data set of 2m surface temperature observations and forecasts for station Hannover}
\format{
An object of class \code{data.frame} with 2191 rows and 30 columns.
}
\source{
Ensemble Forecasts from ECMWF: https://www.ecmwf.int

Observations from DWD: www.dwd.de
}
\usage{
data(station)
}
\description{
Data set of the 2m surface temperature observations, the ensemble mean (mean) and standard
deviation (sd) of various weather quantities for the station
Hannover between 2015-2020. It contains the same variables as the data set
in the references.
}
\examples{
data(station)
}
\references{
Jobst, David, Möller, Annette, & Groß, Jürgen. (2023).
Data set for the ensemble postprocessing of 2m surface temperature forecasts
in Germany for 24 hours lead time (0.1.0) [Data set].
Zenodo. https://doi.org/10.5281/zenodo.8127734
}
\keyword{datasets}
